requirements.txt
Django==4.2.1
djangorestframework==3.14.0
pandas==2.0.2
python-decouple==3.8